requirements.txt
lgpio==0.2.2.0
pad4pi==1.1.5
rpi-lgpio==0.6
